fix: avoid loading fonts from ds to prevent slow loading of module
Fonts loaded from DS bundle are slowing down a lot the loading of the modules
since the request to download the inline font is taking more than 2 seconds to response.

Restore the usage of @fontsource/roboto to load a local copy of the fonts.
Update webpack config to place the fonts where fontsource is loading them from.

refs: SHELL-31