fix: fix bugs
Prevent move of new board to reset size by initializing the internal ref
with the local storage value.

Prevent board position to be calculated when collapsed, since the board container
has size 0 and the calculation will result in a negative value.

Remove listener of the click on mouseUp to always clean it up

refs: SHELL-99