fix: set computed value inside store and not the value received
Value received from the set can be a function, but the store
should contain only the computed value. It should be a clone
of the local storage.

Avoid to set initial value inside store on read if a value is already present.

Add test for checking local storage and store updates.

Fix mock of zustand which was not really resetting the store to
initial value between tests.

Add util to export things for tests only.

Update zustand dependency

refs: SHELL-27