fix: include tinymce translations of help plugin in bundle
refactor: extract the supported locales to a constants

By extracting the locales fixed values to a const object, it is
possible to use this object from the webpack compilation file,
limiting the definition of the supported locales to a single point.
Inside the webpack config, it is now possible to read this object in order
to include inside the bundle only the supported locales for both tinymce
and date-fns.

fix: add missing translations for locales label of settings

chore: update tinymce translations files

Add a field in the supported translations object to map the locale
of i18n to the tinymce one, as already done for date-fns.

refactor: move webpack config to ts

Compile the config before running the build script in order to make it
visible to the sdk.
Add compiled files to .gitignore.
Remove the compiled files after the build.

refs: SHELL-151